+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add Claiming active, Claiming interval fields, Donor information accordion to Template. Refs UIOR-1393.
 * Add ability to hide optional fields in the order template form. Refs UIOR-1394.
 * Disable the "Subscription" checkbox in the PO form when an order workflow status is "Open". Refs UIOR-1426.
+* Add prefix constants and use them in the custom field components. Fixes UIOR-1435.
 
 ## [8.0.3](https://github.com/folio-org/ui-orders/tree/v8.0.2) (2025-04-10)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v8.0.2...v8.0.3)